requirements.txt
nextcord==2.6.0
Pillow~=9.5.0
